Compile on 64-bit MacOSX
*  In preparation to publish, get it to compile on a Mac.
*  The tests fail currently, no doubt as a result of the 64-bit compilation.
   More work needed.

+++ gist-index.cc
@@ -324,7 +324,7 @@ giIndexGist::~giIndexGist()
 
 
 void *
-giIndexGist::operator new(unsigned sz)
+giIndexGist::operator new(size_t sz)
 {
 	giIndexGist * ip = (giIndexGist *)gistInternal::alloc(sz);
 	return (void *)ip;
